Removed VarSpeedServo and fixed Calibration
This update removes VarSpeedServo dependence, opting to use the standard
Servo library instead.  This was due to compiling issues.
Also, the calibration method did not set movement far enough for the x
direction to reach the termination switch if it was fully moved out
along the x direction.  This is now fixed.