Many Improvements
Modified code to take advantage of the newer AccelStepper library that
allows much faster stepping and less blocking when computing steps.
Tuned the max speed and acceleration profiles to minimize travel time
while preserving accuracy.
Modified the code to leave the steppers enabled for 1 second after the
last move so that the position can stabilize as much as possible.  This
is in addition to the precision mode that does not disable the steppers
after movement at all.
Changed calibration to use a slower max speed because of accuracy
problems with a hard stop against the calibration switch.
Made max speed and acceleration variables when can be more easily
modified and potentially updated during operation.